Upload de código-base 5.2.7

+++ basic/source/sbx/sbxdbl.cxx
@@ -198,7 +198,7 @@ start:
         case SbxLPSTR:
             if( !p->pOUString )
                 p->pOUString = new OUString;
-            ImpCvtNum( n, 14, *p->pOUString, bCoreString );
+            ImpCvtNum( (double) n, 14, *p->pOUString, bCoreString );
             break;
         case SbxOBJECT:
         {
@@ -295,7 +295,7 @@ start:
             *p->puInt64 = ImpDoubleToSalUInt64( n ); break;
         case SbxBYREF | SbxDATE:
         case SbxBYREF | SbxDOUBLE:
-            *p->pDouble = n; break;
+            *p->pDouble = (double) n; break;
         case SbxBYREF | SbxCURRENCY:
             if( n > SbxMAXCURR )
             {
